Reduce Catch2 recompilation times
The trick is that there should be **NO TESTS** in the file which compiles the
Catch2 headers: this file should only contain `main`.

Like this, the first compilation of Catch2 in the project is still painfully
slow, but thereafter you can modify the files containing the test specifications
without having to recompile the main file, which would trigger a recompilation
of the headers.