Register and activate sipm_sensitive in constructor
In order for a sensitive detector to work *fully*, you have to jump through (at
least: maybe there are more we haven't found yet) these two hoops:

1. `G4SDManager::GetSDMpointer() -> AddNewDetector(detector);`
2. `detector -> Activate(true)`

Without the second one, it works *partially*. These sorts of pitfalls are a
standard tactic used by the Geant4 developers in their mission to stifle
productivity of their users.

By jumping through the hoops in the constructor, we make it impossible for the
unsuspecting user to forget to do so.

Maybe we should add a nain4 sensitive detector superclass, which will absolve
implementers of sensitive detectors from having to call
`fully_activate_sensitive_detector` in the constructor.