WIP: Looser coupling and explicit dependencies in hdf5/sd/etc.
This is the first version that fully compiles and passes the tests.

Plenty of work left to be done.

There seems to be a problem with the writer: looks like it's not flushing.
Probably because it has no close functionality.

The 'hamamatsu app' test passes if the same instance of `hdf5_io` is used to
read as was used to write. If a different instance is used, the test fails. To
see this, use line 182 instead of 183 in `sipm_hamamatsu_blue-test.cc`. Note:
that line 157 (which deletes the writer) is commented out.