Move SiPM SD in front of active window's optical surface
It seems that the optical surface of the active window prevents optical photons
from entering it, thus none were being registered.

@paolafer gives some historical background:

Apparently this was the case in older versions of Geant4. At some stage, someone
in Nexus noticed that it was no longer necessary: attaching the SD to the active
window itself worked as expected.

As this is a cleaner and more logical organization, lots of Nexus code was
rewritten to take advantage of this simpler scheme.

In Geant4 10.7, the clean, logical approach appears not to work any more, and
thus we need a dummy volume (thickness -> 0) to which we can attach an SD in
front of the optical surface of the volume in which the SD logically belongs.

The need to place the SD in front of the volume that is meant to detect stuff,
is patently absurd, so I hope that this is a bug in G4, or that we're missing
the correct way of doing it.

I have read reports that the corresponding example that comes with G4 stopped
detecting optical photons from 10.6 -> 10.7, which suggests that it *is* a G4 bug.