Fix bug in positioning of pre-active_window
Two mysterious test crashes remain. These are both segfaults at the end of main
when in a chain of destructors:

RunManager -> RunManagerKernel -> SDManager -> SDStructure

So they occur long after the tests have passed and finished.

They appear when the pre-active window is present, and disappear when the
pre-active window is removed.