Demonstrate that steel absorbs rather than reflecting
The optical photons TIR off the quartz, inside the xenon. We weren't sure if
this was cause by erroneous reflection off steel. Here we demonstrate that the
steel does indeed absorb the optical photons in the xenon, by

+ Adding a layer of steel outside the xenon, before the quartz
+ Adding two parallel strips of steel inside the xenon.

By visualising the first event, it is clear that the lower steel strip
does not reflect any optical photons. The two strips are just to the right of
centre at the top of the xenon ring.