Buffer HDF5 writing and double the speed!
We were still using the first prototype of HighFive use. As I
suspected (and commented in the code), it was hopelessly wasteful and
slow.

Profiling showed that HighFive was taking up close to 50% of the
simulation time.

By buffering the data before sending them to HighFive, we reduce the
time spent in HighFive to essentially zero.

VoxelNavigation now takes 67% of the time ... another long-term
suspect (given how slow attenuation map generation is in complex
geometries).