Switch Jaszcak sources from PMMA to water
... on suspicion of the corresponding density map (with source density
19% in excess of background density) making the attenuation correction
pick out the sources, and giving images which are too good to be true:
the density map identifies the location of the sources with perfect
accuracy, consequently attenuation correction emphasizes sources and
de-emphasizes background.

It reality we these should be thin PMMA shells containing water, but
this approximation (water held in place by magic) should do for now.